+++ src/test/scala/edu/ie3/mobsim/utils/IoUtilsSpec.scala
@@ -16,6 +16,7 @@ import edu.ie3.util.quantities.PowerSystemUnits.KILOWATTHOUR
 import java.io.{BufferedReader, File, FileReader}
 import java.nio.file.Path
 import java.util
+import java.util.UUID
 
 class IoUtilsSpec extends UnitSpec with IoUtilsTestData {
   "IoUtils" should {
@@ -129,7 +130,7 @@ class IoUtilsSpec extends UnitSpec with IoUtilsTestData {
     }
 
     "write pois correctly" in {
-      ioUtils.writePois(poiMap)
+      ioUtils.writePois(poiMap, Map.empty[UUID, UUID])
 
       val data = new BufferedReader(
         new FileReader(new File(outputFileDir, "pois.csv"))